codecloud/requirements.txt
fastapi===0.111.0
python-dotenv===1.0.1
mysql-connector-python===9.0.0